refactor: move timestamp converter to separate file
