+++ src/components/ChatPreview/ChatPreview.tsx
@@ -3,13 +3,13 @@ import { useContext, useState } from 'react';
 import {
   onSnapshot, doc, updateDoc,
 } from 'firebase/firestore';
-import * as firestore from 'firebase/firestore';
 import { db } from '../../firebaseConfig';
 import Avatar from '../Avatar/Avatar';
 import { AuthContext } from '../../context/AuthContext';
 import { ActiveChatContext } from '../../context/ActiveChatContext';
 import type { User, UserChat } from '../../types';
 import './ChatPreview.scss';
+import { convertTimestamp } from '../../hooks/timestampConverter';
 
 interface ChatPreviewProps {
   data: UserChat,
@@ -54,22 +54,6 @@ function ChatPreview({
       setIsOnlineStatus(userData.isOnline);
     });
   }
-
-  const convertTimestamp = (timestamp: firestore.Timestamp): string => {
-    const currentDate = timestamp.toDate();
-    const now = new Date();
-
-    if (currentDate.getFullYear() === now.getFullYear()
-      && currentDate.getMonth() === now.getMonth()
-      && currentDate.getDate() === now.getDate()) {
-      const hours = currentDate.getHours().toString().padStart(2, '0');
-      const minutes = currentDate.getMinutes().toString().padStart(2, '0');
-      return `${hours}:${minutes}`;
-    }
-    const month = currentDate.toLocaleString('default', { month: 'short' });
-    const day = currentDate.getDate();
-    return `${month} ${day}`;
-  };
 
   const handleContextMenu = (event: React.MouseEvent) => {
     onContextMenu(event, uid);
